Update : #71 과도한 state 끌어올리기 수정하기
- Edit 과정에서 UI 버그가 발생해서 살펴보니 state의 과도한 끌어올림을 발견해서 수정
