+++ src/molecules/TodoCard/index.tsx
@@ -1,9 +1,5 @@
-import { ReactEventHandler, useState } from 'react';
-
 import EditUI from './content/EditUI';
 import TodoUI from './content/TodoUI';
-
-import { categoryValidation } from '../../shared/inputValidation';
 
 import { useEdit } from '../../hooks';
 
@@ -24,6 +20,7 @@ interface ITodoCardProps {
 }
 
 const TodoCard = ({ todoData, dragHandleProps, snapshot }: ITodoCardProps) => {
+  const { id } = todoData;
   const queryClient = useQueryClient();
 
   // useMutation 요거 구현하기
@@ -39,78 +36,15 @@ const TodoCard = ({ todoData, dragHandleProps, snapshot }: ITodoCardProps) => {
     },
   });
 
-  const { id, todo: todoTitle, categories, duration: tomato } = todoData;
-
   const [{ editMode, editTodoId }, setIsEdit] = useEdit();
-
-  const [showEdit, setShowEdit] = useState(false);
-  const [titleValue, setTitleValue] = useState(todoTitle);
-  const [categoryArray, setCategoryArray] = useState(categories);
-  const [categoryValue, setCategoryValue] = useState(''); // 새로운 카테고리에 대한 input value state
-  const [duration, setDuration] = useState(tomato);
-
-  const handleMouseOver = () => {
-    setShowEdit(true);
-  };
-
-  const handleMouseOut = () => {
-    setShowEdit(false);
-  };
 
   const handleEditButton = () => {
     setIsEdit({ editMode: true, editTodoId: id });
   };
 
-  const handleChangeTitle = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setTitleValue(event.target.value);
-  };
-
-  const handleChangeCategory = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setCategoryValue(event.target.value);
-  };
-
-  const handleCategorySubmit = (
-    event: React.KeyboardEvent<HTMLInputElement>,
-  ) => {
-    if (event.code === 'Enter') {
-      // 한글 중복 입력 처리
-      if (event.nativeEvent.isComposing) return;
-
-      const newCategory = (event.target as HTMLInputElement).value;
-
-      const trimmed = categoryValidation(newCategory, categoryArray ?? []);
-
-      if (!trimmed) return;
-
-      if (categoryArray) {
-        const copy = categoryArray.slice();
-        copy.push(trimmed);
-
-        setCategoryArray(copy);
-      } else {
-        setCategoryArray([trimmed]);
-      }
-
-      setCategoryValue('');
-    }
-  };
-
-  const handleDeleteCategory = (category: string) => {
-    setCategoryArray((prev) => {
-      const deleted = prev?.filter((tag) => tag !== category) as string[]; // QUESTION event.currentTarget.innerHTML를 바로 넣어주면 에러가 왜 날까?
-      return deleted;
-    });
-  };
-
-  const handleDuration: ReactEventHandler<HTMLSelectElement> = (event) => {
-    setDuration(Number(event.currentTarget.value));
   };
 
   const handleEditCancel = () => {
-    setIsEdit({ editMode: false, editTodoId: undefined });
-  };
-  const handleEditSubmit = (todo: TodoEntity) => {
-    mutate(todo);
     setIsEdit({ editMode: false, editTodoId: undefined });
   };
 
@@ -120,17 +54,8 @@ const TodoCard = ({ todoData, dragHandleProps, snapshot }: ITodoCardProps) => {
         return (
           <EditUI
             todoData={todoData}
-            title={titleValue}
-            handleChangeTitle={handleChangeTitle}
-            category={categoryValue}
-            handleChangeCategory={handleChangeCategory}
-            handleClickCategory={handleDeleteCategory}
-            handleCategorySubmit={handleCategorySubmit}
-            categories={categoryArray}
             handleEditCancel={handleEditCancel}
             handleEditSubmit={handleEditSubmit}
-            duration={duration}
-            handleDuration={handleDuration}
           />
         );
       case false:
@@ -140,9 +65,6 @@ const TodoCard = ({ todoData, dragHandleProps, snapshot }: ITodoCardProps) => {
             snapshot={snapshot}
             todoData={todoData}
             editMode={editMode}
-            showEdit={showEdit}
-            handleMouseOut={handleMouseOut}
-            handleMouseOver={handleMouseOver}
             handleEditButton={handleEditButton}
           />
         );
